requirements.txt
pynput==1.6.8
requests==2.24.0
Pillow==7.2.0
PyYAML==5.3.1
